updated docstrings, moved encoders to base file
